fix warning.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@810 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/languages/mod_spidermonkey/mod_spidermonkey.c
@@ -721,7 +721,7 @@ static size_t hash_callback(void *ptr, size_t size, size_t nmemb, void *data)
 
 static size_t file_callback(void *ptr, size_t size, size_t nmemb, void *data)
 {
-	register size_t realsize = size * nmemb;
+	register unsigned int realsize = (unsigned int)(size * nmemb);
 	struct config_data *config_data = data;
 
 	write(config_data->fd, ptr, realsize);
